airflow/requirements.txt
apache-airflow==3.1.5
yfinance==1.0
scikit-learn==1.8.0
pandas==2.3.3
psycopg2-binary==2.9.11
sec-edgar-downloader==5.1.0
beautifulsoup4==4.14.3
python-dotenv==1.2.1
langchain-text-splitters==1.1.1
langchain-openai==1.1.10
pgvector==0.4.2